Renamed fanout sidecar to in-memory dispatcher
