--- cmd/fanoutsidecar/kodata/LICENSE
@@ -1,1 +1,0 @@
-../../../LICENSE
