--- cmd/fanoutsidecar/kodata/VENDOR-LICENSE
@@ -1,1 +1,0 @@
-../../../third_party/VENDOR-LICENSE
