+++ config/provisioners/in-memory-channel/in-memory-channel.yaml
@@ -193,11 +193,7 @@ spec:
       serviceAccountName: in-memory-channel-dispatcher
       containers:
         - name: dispatcher
-          image: github.com/knative/eventing/cmd/fanoutsidecar
-          args:
-            - --sidecar_port=8080
-            - --channel_provisioner=in-memory
-            - --channel_provisioner=in-memory-channel
+          image: github.com/knative/eventing/pkg/provisioners/inmemory/dispatcher
           env:
           - name: SYSTEM_NAMESPACE
             valueFrom:
