+++ contrib/kafka/pkg/dispatcher/dispatcher.go
@@ -29,8 +29,8 @@ import (
 	"github.com/knative/eventing/contrib/kafka/pkg/controller"
 	eventingduck "github.com/knative/eventing/pkg/apis/duck/v1alpha1"
 	"github.com/knative/eventing/pkg/provisioners"
+	"github.com/knative/eventing/pkg/provisioners/multichannelfanout"
 	topicUtils "github.com/knative/eventing/pkg/provisioners/utils"
-	"github.com/knative/eventing/pkg/sidecar/multichannelfanout"
 )
 
 type KafkaDispatcher struct {
