+++ contrib/kafka/pkg/dispatcher/dispatcher_test.go
@@ -20,8 +20,8 @@ import (
 
 	eventingduck "github.com/knative/eventing/pkg/apis/duck/v1alpha1"
 	"github.com/knative/eventing/pkg/provisioners"
-	"github.com/knative/eventing/pkg/sidecar/fanout"
-	"github.com/knative/eventing/pkg/sidecar/multichannelfanout"
+	"github.com/knative/eventing/pkg/provisioners/fanout"
+	"github.com/knative/eventing/pkg/provisioners/multichannelfanout"
 	_ "github.com/knative/pkg/system/testing"
 )
