+++ pkg/provisioners/fanout/fanout_handler.go
