+++ pkg/provisioners/multichannelfanout/config.go
@@ -18,7 +18,7 @@ package multichannelfanout
 
 import (
 	"github.com/knative/eventing/pkg/apis/eventing/v1alpha1"
-	"github.com/knative/eventing/pkg/sidecar/fanout"
+	"github.com/knative/eventing/pkg/provisioners/fanout"
 )
 
 // Config for a multichannelfanout.Handler.
