+++ pkg/provisioners/multichannelfanout/multi_channel_fanout_handler_test.go
@@ -25,7 +25,7 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 	eventingduck "github.com/knative/eventing/pkg/apis/duck/v1alpha1"
-	"github.com/knative/eventing/pkg/sidecar/fanout"
+	"github.com/knative/eventing/pkg/provisioners/fanout"
 	"go.uber.org/zap"
 )
